pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "polars>=1.33.1,<2.0",
    "pydantic>=2.11.9,<3.0",
    "pydantic-settings>=2.10.1,<3.0",
]

[dependency-groups]
dev = [
    "apache-airflow==3.0.6",
    "mypy==1.18.2",
    "pandas==2.3.2",
    "pre-commit==4.3.0",
    "pyarrow==21.0.0",
    "pytest==8.4.2",
    "ruff==0.13.1",
    "types-requests==2.32.4.20250913",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "scripts",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
warn_return_any = true
warn_unused_ignores = true
strict_equality = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
